// Qanat.Web/src/app/shared/generated/model/fuel-type-simple-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class FuelTypeSimpleDto { 
    FuelTypeID?: number;
    FuelTypeName?: string;
    FuelTypeDisplayName?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface FuelTypeSimpleDtoForm { 
    FuelTypeID?: FormControl<number>;
    FuelTypeName?: FormControl<string>;
    FuelTypeDisplayName?: FormControl<string>;
}

export class FuelTypeSimpleDtoFormControls { 
    public static FuelTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FuelTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FuelTypeDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}