// Qanat.Web/src/app/shared/generated/model/well-registration-detailed-dto.ts
export class WellRegistrationDetailedDto { 
    WellRegistrationID?: number;
    GeographyID?: number;
    Geography?: GeographyDisplayDto;
    WellID?: number;
    WellName?: string;
    WellStatusRegistrationID?: number;
    WellRegistrationStatus?: WellRegistrationStatusSimpleDto;
    ParcelID?: number;
    Parcel?: ParcelDisplayDto;
    StateWCRNumber?: string;
    CountyWellPermitNumber?: string;
    DateDrilled?: string;
    WellDepth?: number;
    SubmitDate?: string;
    ApprovalDate?: string;
    CreateUserGuid?: string;
    CreateUserID?: number;
    CreateUserEmail?: string;
    FairyshrimpWellID?: number;
    Latitude?: number;
    Longitude?: number;
    IrrigatedParcels?: Array<ParcelDisplayDto>;
    LandownerContact?: WellRegistrationContactWithStateDto;
    OwnerOperatorContact?: WellRegistrationContactWithStateDto;
    WellRegistrationMetadatum?: WellRegistrationMetadatumSimpleDto;
    WellRegistrationWaterUses?: Array<WellRegistrationWaterUseDisplayDto>;
    WellRegistrationFileResources?: Array<WellRegistrationFileResourceDto>;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}